+++ src/main/java/com/movies/entity/Movies.java
@@ -10,32 +10,35 @@ public class Movies extends AuditableEntity {
 
 	@Column(name = "title")
 	private String title;
-	
+
 	@Column(name = "category")
 	private String category;
-	
+
 	@Column(name = "starRating")
-	private Long starRating;
-	
-	
+	private Float starRating;
+
 	public String getTitle() {
 		return title;
 	}
+
 	public void setTitle(String title) {
 		this.title = title;
 	}
+
 	public String getCategory() {
 		return category;
 	}
+
 	public void setCategory(String category) {
 		this.category = category;
 	}
-	public Long getStarRating() {
+
+	public Float getStarRating() {
 		return starRating;
 	}
-	public void setStarRating(Long starRating) {
+
+	public void setStarRating(Float starRating) {
 		this.starRating = starRating;
 	}
-	
-	
+
 }
